[PROC] Major addition of P-ROC input logic.  This allows certain switch events from the P-ROC to affect the UI.  Specifically, showcopyright() and displaygameinfo() can now be acknowledge by P-ROC switches.  Also, the specific P-ROC button combination of both flippers and the start button will now exit the simulation.  Additionally, the showcopyright() and displaygameinfo() functions will now drive text to the P-ROC's dmd or alphanumeric display to instruct the user.
Significant reorganization of P-ROC initialization logic so the P-ROC can be partially initialized before the simulation is running.  This is done to allow the displays and switches to be used as described above for the user interface.

Adjusted P-ROC alphanumeric DMD segement maps for cleaner text.

Only update the P-ROC's DMD buffers when a P-ROC DMD event is received.

Enable g_fHandleKeyboard and g_fHandleMechanics by default.  Only disable if the P-ROC is up.